revert refactor on @dcl/sdk
