+++ .eslintignore
@@ -3,6 +3,9 @@ packages/@dcl/js-runtime
 *.d.ts
 test/sdk/simple-scene
 packages/@dcl/ecs/dist/
+packages/@dcl/sdk/dist/
+packages/@dcl/sdk/internal/
+packages/@dcl/sdk/testing/
 packages/@dcl/playground-assets/dist
 packages/@dcl/ecs/src/components/generated
 tmp
