+++ packages/@dcl/sdk/src/with-composite.ts
