+++ test/ecs/components/components.spec.ts
@@ -1,6 +1,6 @@
 import { Engine, Entity, components } from '../../../packages/@dcl/ecs/src'
 import { ReadWriteByteBuffer } from '../../../packages/@dcl/ecs/src/serialization/ByteBuffer'
-import { Quaternion, Vector3 } from '../../../packages/@dcl/sdk/math'
+import { Quaternion, Vector3 } from '../../../packages/@dcl/sdk/src/math'
 
 describe('Legacy component tests', () => {
   it('cube example scene', async () => {
