+++ test/ecs/ecs.crdt.spec.ts
@@ -9,7 +9,7 @@ import {
 import { PutComponentOperation } from '../../packages/@dcl/ecs/src/serialization/crdt'
 import { Entity, EntityState, EntityUtils, RESERVED_STATIC_ENTITIES } from '../../packages/@dcl/ecs/src/engine/entity'
 import { ReadWriteByteBuffer } from '../../packages/@dcl/ecs/src/serialization/ByteBuffer'
-import { Vector3 } from '../../packages/@dcl/sdk/math'
+import { Vector3 } from '../../packages/@dcl/sdk/src/math'
 import { SandBox, wait } from './utils'
 
 async function simpleScene(engine: IEngine) {
