+++ test/ecs/math.spec.ts
@@ -1,4 +1,4 @@
-import * as Math from '../../packages/@dcl/sdk/math'
+import * as Math from '../../packages/@dcl/sdk/src/math'
 
 describe('Events System', () => {
   it('should get Math exported fns', () => {
